pyproject.toml
[project]
name = "gazzette-api"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.7.2",
    "api-analytics>=1.2.4",
    "asyncpg>=0.29.0",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.112.1",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.27.0",
    "jinja2>=3.1.4",
    "procrastinate>=2.0.0",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.2.0",
    "python-decouple>=3.8",
    "redis>=5.0.8",
    "requests>=2.32.3",
    "tortoise-orm>=0.21.6",
    "uvicorn[standard]>=0.30.6",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.2",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88
